% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rwards-package.R
\docType{package}
\name{Rwards-package}
\alias{Rwards}
\alias{Rwards-package}
\title{Rwards: Reward Users for Discovering Bugs}
\description{
Transforms the learning experience of a programming language by rewarding users for making mistakes. The goal is to take the anxiety out of error messages and make learning programming more engaging for newcomers. Users earn points for each error encountered, encouraging exploration and experimentation without fear of failure. Progress is cached across sessions.
}
\details{
Attaching the package with \code{library(Rwards)} installs an error handler that
turns each error into experience points and levels, shown with a progress
bar and an encouraging message. Detaching the package restores the error
handler that was in place beforehand.
}
\seealso{
\code{\link[=evaluate_code]{evaluate_code()}}
}
\author{
\strong{Maintainer}: Mickael Temporão \email{m.temporao@sciencespobordeaux.fr} [copyright holder]

Authors:
\itemize{
  \item Mickael Temporão \email{m.temporao@sciencespobordeaux.fr} [copyright holder]
  \item Alexandre Fortier-Chouinard
}

}
\keyword{internal}
