% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rwards-package.R
\docType{package}
\name{Rwards-package}
\alias{Rwards}
\alias{Rwards-package}
\title{Rwards: Turn R Errors into Experience Points and Levels}
\description{
Turns R errors into a game. Each error earns experience points, with a bonus the first time a given error is seen, and accumulated points raise the user through titled levels shown with a progress bar and an encouraging message. Progress is cached across sessions. The goal is to take the anxiety out of error messages and make learning R more engaging for newcomers.
}
\details{
Attaching the package with \code{library(Rwards)} installs an error handler that
turns each error into experience points and levels, shown with a progress
bar and an encouraging message. Detaching the package restores the error
handler that was in place beforehand.
}
\seealso{
\code{\link[=evaluate_code]{evaluate_code()}}
}
\author{
\strong{Maintainer}: Mickael Temporão \email{m.temporao@sciencespobordeaux.fr} [copyright holder]

Authors:
\itemize{
  \item Mickael Temporão \email{m.temporao@sciencespobordeaux.fr} [copyright holder]
  \item Alexandre Fortier-Chouinard
}

}
\keyword{internal}
