\name{Rwards-package}
\alias{Rwards-package}
\alias{Rwards}
\docType{package}
\title{\packageTitle{Rwards}}
\description{
\packageDescription{Rwards}
}
\details{
Attaching the package with \code{library(Rwards)} installs an error handler that
turns each error into experience points and levels, shown with a progress bar
and an encouraging message. Detaching the package restores the error handler
that was in place beforehand.
}
\author{
\packageAuthor{Rwards}

Maintainer: \packageMaintainer{Rwards}
}
\seealso{
\code{\link{evaluate_code}}
}
\keyword{package}
