\name{Rwards-package}
\alias{Rwards-package}
\alias{Rwards}
\docType{package}
\title{
\packageTitle{Rwards}
}
\description{
\packageDescription{Rwards}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{Rwards}

Maintainer: \packageMaintainer{Rwards}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = Rwards)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{Rwards}}
%\section{Documentation Index}{\packageIndices{Rwards}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
